+++ crates/rust-client/src/store/mod.rs
@@ -68,15 +68,9 @@ mod smt_forest;
 pub use smt_forest::AccountSmtForest;
 
 mod account;
-pub use account::{
-    AccountRecord,
-    AccountRecordData,
-    AccountStatus,
-    AccountUpdates,
-    ClientAccountType,
-};
+pub use account::{AccountRecord, AccountRecordData, AccountStatus, ClientAccountType};
 
-pub use crate::sync::PublicAccountUpdate;
+pub use crate::sync::{AccountUpdates, PublicAccountUpdate};
 mod note_record;
 pub use note_record::{
     InputNoteRecord,